+++ server/src/enums/Filter.ts
@@ -1,0 +1,4 @@
+export enum FilterType {
+    List = 'list',
+    Range = 'range',
+}
