+++ server/src/enums/Label.ts
@@ -1,0 +1,5 @@
+export enum LabelType {
+    DISCOUNT = 'discount',
+    NEW = 'new',
+    POPULAR = 'popular'
+}
