+++ server/src/enums/Order.ts
@@ -1,0 +1,7 @@
+export enum OrderStatus {
+    INIT = 1,
+    POSTED,
+    IN_PROCESS,
+    COMPLETED,
+    CANCELED
+}
