+++ server/src/enums/Product.ts
@@ -1,0 +1,4 @@
+export enum ProductStatus {
+    AVAILABLE = 1,  // в наличии
+    OUT_OF_STOCK    // распродан
+}
