+++ server/src/enums/Query.ts
@@ -1,0 +1,6 @@
+export enum SortType {
+    POPULAR = 1,  // most popular first
+    PRICE_UP,     // from cheap to expensive
+    PRICE_DOWN ,  // from expensive to cheap
+    NEW           // most new first
+}
