+++ server/src/interfaces/Product.ts
@@ -1,11 +1,8 @@
 import { ICategory } from '@interfaces/Category';
 import { IImage } from '@interfaces/Image';
 import { IProperty } from '@interfaces/Property';
-
-export enum ProductStatus {
-    AVAILABLE = 1,  // в наличии
-    OUT_OF_STOCK    // распродан
-}
+import { ILabel } from '@interfaces/Label';
+import { ProductStatus } from '@enums/Product';
 
 export interface IPrice {
     current: number
@@ -17,6 +14,7 @@ export interface IBaseProduct {
     title: string
     status: ProductStatus
     price?: IPrice
+    labels?: Array<ILabel>
 }
 
 export interface IProductPreview extends IBaseProduct {
