+++ server/src/interfaces/Queries.ts
@@ -1,11 +1,5 @@
-import { ProductStatus } from './Product';
-
-export enum SortType {
-    POPULAR = 1,  // most popular first
-    PRICE_UP,     // from cheap to expensive
-    PRICE_DOWN ,  // from expensive to cheap
-    NEW           // most new first
-}
+import { ProductStatus } from '@enums/Product';
+import { SortType } from '@enums/Query';
 
 export interface IPriceSearchQuery {
     min: number,
