+++ server/src/services/Product.ts
@@ -7,7 +7,8 @@ import { ProductEntity } from '@entities/Product';
 import { FOLDER_TYPES, FSService } from '@services/FS';
 import { ImageService } from '@services/Image';
 import { ICookies } from '@interfaces/Cookies';
-import { ISearchReqQueries, SortType } from '@interfaces/Queries';
+import { ISearchReqQueries } from '@interfaces/Queries';
+import { SortType } from '@enums/Query';
 import { SearchQueriesDTO } from '@dto/Queries';
 import { PaginationResultDTO } from '@dto/Pagination';
 import { CommonDashboardsDTO, UserDashboardsDTO } from '@dto/Dashboard';
